+++ build.gradle
@@ -1,9 +1,9 @@
 buildscript {
-    ext.kotlin_version = '1.1.2-5'
+    ext.kotlin_version = '1.1.3'
 
     repositories {
         jcenter()
-        maven { url "https://maven.google.com" }
+        google()
     }
 
     dependencies {
@@ -16,6 +16,6 @@ buildscript {
 allprojects {
     repositories {
         jcenter()
-        maven { url "https://maven.google.com" }
+        google()
     }
 }
