+++ demo/src/main/java/com/devbrackets/android/playlistcoredemo/ui/activity/AudioPlayerActivity.java
@@ -16,7 +16,7 @@ import com.devbrackets.android.playlistcore.event.MediaProgress;
 import com.devbrackets.android.playlistcore.event.PlaylistItemChange;
 import com.devbrackets.android.playlistcore.listener.PlaylistListener;
 import com.devbrackets.android.playlistcore.listener.ProgressListener;
-import com.devbrackets.android.playlistcore.service.BasePlaylistService;
+import com.devbrackets.android.playlistcore.service.PlaybackState;
 import com.devbrackets.android.playlistcoredemo.App;
 import com.devbrackets.android.playlistcoredemo.R;
 import com.devbrackets.android.playlistcoredemo.data.MediaItem;
@@ -105,7 +105,7 @@ public class AudioPlayerActivity extends AppCompatActivity implements PlaylistLi
     }
 
     @Override
-    public boolean onPlaybackStateChanged(@NonNull BasePlaylistService.PlaybackState playbackState) {
+    public boolean onPlaybackStateChanged(@NonNull PlaybackState playbackState) {
         switch (playbackState) {
             case STOPPED:
                 finish();
@@ -157,8 +157,8 @@ public class AudioPlayerActivity extends AppCompatActivity implements PlaylistLi
             onPlaylistItemChanged(itemChange.getCurrentItem(), itemChange.getHasNext(), itemChange.getHasPrevious());
         }
 
-        BasePlaylistService.PlaybackState currentPlaybackState = playlistManager.getCurrentPlaybackState();
-        if (currentPlaybackState != BasePlaylistService.PlaybackState.STOPPED) {
+        PlaybackState currentPlaybackState = playlistManager.getCurrentPlaybackState();
+        if (currentPlaybackState != PlaybackState.STOPPED) {
             onPlaybackStateChanged(currentPlaybackState);
         }
